requirements.txt
pytorch_lightning
transformers
Korpora
wandb
typing-extensions
datasets
omegaconf
accelerate
wget
pandas
numpy
matplotlib
python-dotenv
